+++ test/test_helper.exs
@@ -1,5 +1,6 @@
 ExUnit.start
 
+{:ok, _} = Application.ensure_all_started(:hackney)
 Logger.configure_backend(:console, colors: [enabled: false], metadata: [:request_id])
 
 defmodule Plug.ProcessStore do